Refactor DonationScreen
