+++ app/src/main/java/ie/setu/project/views/donation/DonationScreen.kt
@@ -1,35 +1,19 @@
 package ie.setu.project.views.donation
 
-import android.net.Uri
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.*
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import coil.compose.AsyncImage
 import ie.setu.project.R
-import ie.setu.project.models.clothing.ClosetOrganiserModel
-import java.text.SimpleDateFormat
-import java.util.*
-import java.util.concurrent.TimeUnit
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -40,18 +24,25 @@ fun DonationScreen(
     val flaggedItems    by vm.flaggedItems.collectAsStateWithLifecycle()
     val thresholdMonths by vm.thresholdMonths.collectAsStateWithLifecycle()
     val remindersOn     by vm.remindersEnabled.collectAsStateWithLifecycle()
-    val donatedCount    by vm.donatedCount.collectAsStateWithLifecycle()
+    val donationStats   by vm.donationStats.collectAsStateWithLifecycle()
+    val pendingPlans    by vm.pendingPlans.collectAsStateWithLifecycle()
+    val selectedItem    by vm.selectedItem.collectAsStateWithLifecycle()
 
     var showSettings by remember { mutableStateOf(false) }
+    var activeTab    by remember { mutableStateOf(0) }
 
     LaunchedEffect(Unit) { vm.refreshFlaggedItems() }
+
+    if (selectedItem != null) {
+        DonatePlanSheet(item = selectedItem!!, vm = vm, onDismiss = { vm.clearSelectedItem() })
+    }
 
     Scaffold(
         topBar = {
             TopAppBar(
                 navigationIcon = {
                     IconButton(onClick = onBack) {
-                        Icon(Icons.Default.ArrowBack, contentDescription = "Back", tint = Color.White)
+                        Icon(Icons.Default.ArrowBack, "Back", tint = Color.White)
                     }
                 },
                 title = {
@@ -60,70 +51,61 @@ fun DonationScreen(
                         horizontalArrangement = Arrangement.Center,
                         modifier = Modifier.fillMaxWidth()
                     ) {
-                        Icon(painter = painterResource(id = R.drawable.ic_heart), contentDescription = null, tint = Color.White, modifier = Modifier.size(24.dp))
-                        Spacer(modifier = Modifier.width(8.dp))
-                        Text("Donation Tracker", fontSize = 30.sp, fontFamily = androidx.compose.ui.text.font.FontFamily.Cursive, color = Color.White)
-                        Spacer(modifier = Modifier.width(8.dp))
-                        Icon(painter = painterResource(id = R.drawable.ic_heart), contentDescription = null, tint = Color.White, modifier = Modifier.size(24.dp))
+                        Icon(painterResource(R.drawable.ic_heart), null, tint = Color.White,
+                            modifier = Modifier.size(22.dp))
+                        Spacer(Modifier.width(8.dp))
+                        Text("Donation Tracker", fontSize = 28.sp,
+                            fontFamily = androidx.compose.ui.text.font.FontFamily.Cursive,
+                            color = Color.White)
+                        Spacer(Modifier.width(8.dp))
+                        Icon(painterResource(R.drawable.ic_heart), null, tint = Color.White,
+                            modifier = Modifier.size(22.dp))
                     }
                 },
                 actions = {
                     IconButton(onClick = { showSettings = !showSettings }) {
-                        Icon(Icons.Default.Settings, contentDescription = "Settings", tint = Color.White)
+                        Icon(Icons.Default.Settings, "Settings", tint = Color.White)
                     }
                 },
                 colors = TopAppBarDefaults.topAppBarColors(containerColor = MaterialTheme.colorScheme.primary)
             )
         }
     ) { padding ->
-
         Column(modifier = Modifier.padding(padding).fillMaxSize()) {
 
             if (showSettings) {
                 DonationSettingsPanel(
-                    thresholdMonths  = thresholdMonths,
-                    remindersEnabled = remindersOn,
+                    thresholdMonths   = thresholdMonths,
+                    remindersEnabled  = remindersOn,
                     onThresholdChange = { vm.setThresholdMonths(it) },
                     onRemindersChange = { vm.setRemindersEnabled(it) }
                 )
             }
 
-
             Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .background(MaterialTheme.colorScheme.primaryContainer)
-                    .padding(horizontal = 20.dp, vertical = 12.dp),
+                modifier = Modifier.fillMaxWidth().padding(horizontal = 20.dp, vertical = 12.dp),
                 horizontalArrangement = Arrangement.SpaceBetween,
                 verticalAlignment = Alignment.CenterVertically
             ) {
-                SummaryColumn("${flaggedItems.size}", "Flagged",   MaterialTheme.colorScheme.primary)
-                SummaryColumn("$donatedCount",        "Donated",   Color(0xFF2E7D32))
-                SummaryColumn("${thresholdMonths}mo", "Threshold", MaterialTheme.colorScheme.tertiary)
+                SummaryColumn("${flaggedItems.size}",         "Flagged",   MaterialTheme.colorScheme.primary)
+                SummaryColumn("${donationStats.totalDonated}","Donated",   Color(0xFF2E7D32))
+                SummaryColumn("${pendingPlans.size}",         "Planned",   Color(0xFFF57C00))
+                SummaryColumn("${thresholdMonths}mo",         "Threshold", MaterialTheme.colorScheme.tertiary)
             }
 
-            if (flaggedItems.isEmpty()) {
-                EmptyDonationState()
-            } else {
-                Text(
-                    "Items unworn for ${thresholdMonths}+ months",
-                    modifier = Modifier.padding(start = 16.dp, top = 16.dp, bottom = 4.dp),
-                    style = MaterialTheme.typography.titleMedium,
-                    color = MaterialTheme.colorScheme.primary,
-                    fontWeight = FontWeight.Bold
-                )
-                LazyColumn(
-                    contentPadding = PaddingValues(horizontal = 16.dp, vertical = 8.dp),
-                    verticalArrangement = Arrangement.spacedBy(10.dp)
-                ) {
-                    items(flaggedItems, key = { it.id }) { item ->
-                        DonationItemCard(
-                            item    = item,
-                            onDonate = { vm.markForDonation(item) },
-                            onKeep   = { vm.keepItem(item) }
-                        )
-                    }
-                }
+            TabRow(selectedTabIndex = activeTab) {
+                Tab(selected = activeTab == 0, onClick = { activeTab = 0 },
+                    text = { Text("To Donate") }, icon = { Icon(Icons.Default.Checkroom, null) })
+                Tab(selected = activeTab == 1, onClick = { activeTab = 1 },
+                    text = { Text("Planned") },   icon = { Icon(Icons.Default.Event, null) })
+                Tab(selected = activeTab == 2, onClick = { activeTab = 2 },
+                    text = { Text("Stats") },     icon = { Icon(Icons.Default.BarChart, null) })
+            }
+
+            when (activeTab) {
+                0 -> FlaggedTab(flaggedItems, thresholdMonths, onDonate = { vm.startDonationFlow(it) }, onKeep = { vm.keepItem(it) })
+                1 -> PlannedTab(pendingPlans, onConfirm = { vm.confirmDonation(it) }, onCancel = { vm.cancelPlan(it) })
+                2 -> StatsTab(donationStats)
             }
         }
     }
@@ -132,7 +114,7 @@ fun DonationScreen(
 @Composable
 private fun SummaryColumn(value: String, label: String, color: Color) {
     Column(horizontalAlignment = Alignment.CenterHorizontally) {
-        Text(value, fontSize = 28.sp, fontWeight = FontWeight.Bold, color = color)
+        Text(value, fontSize = 28.sp, fontWeight = androidx.compose.ui.text.font.FontWeight.Bold, color = color)
         Text(label, fontSize = 12.sp, color = MaterialTheme.colorScheme.onPrimaryContainer)
     }
 }
@@ -145,14 +127,13 @@ private fun DonationSettingsPanel(
     onRemindersChange: (Boolean) -> Unit
 ) {
     val options = listOf(3, 6, 9, 12)
-
     Card(
         modifier = Modifier.fillMaxWidth().padding(12.dp),
-        shape = RoundedCornerShape(14.dp),
+        shape = androidx.compose.foundation.shape.RoundedCornerShape(14.dp),
         colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surfaceVariant)
     ) {
         Column(modifier = Modifier.padding(16.dp), verticalArrangement = Arrangement.spacedBy(12.dp)) {
-            Text("Inactivity Threshold", fontWeight = FontWeight.SemiBold, fontSize = 15.sp)
+            Text("Inactivity Threshold", fontWeight = androidx.compose.ui.text.font.FontWeight.SemiBold, fontSize = 15.sp)
             Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
                 options.forEach { months ->
                     FilterChip(
@@ -166,14 +147,14 @@ private fun DonationSettingsPanel(
                     )
                 }
             }
-            Divider()
+            HorizontalDivider()
             Row(
                 modifier = Modifier.fillMaxWidth(),
                 horizontalArrangement = Arrangement.SpaceBetween,
                 verticalAlignment = Alignment.CenterVertically
             ) {
                 Column {
-                    Text("Donation Reminders", fontWeight = FontWeight.SemiBold, fontSize = 15.sp)
+                    Text("Donation Reminders", fontWeight = androidx.compose.ui.text.font.FontWeight.SemiBold, fontSize = 15.sp)
                     Text("Show badge on nav bar", fontSize = 12.sp, color = Color.Gray)
                 }
                 Switch(
@@ -185,113 +166,6 @@ private fun DonationSettingsPanel(
                     )
                 )
             }
-        }
-    }
-}
-
-@Composable
-fun DonationItemCard(
-    item: ClosetOrganiserModel,
-    onDonate: () -> Unit,
-    onKeep: () -> Unit
-) {
-    val daysSince = TimeUnit.MILLISECONDS.toDays(Date().time - item.lastWorn.time)
-    val sdf = remember { SimpleDateFormat("dd MMM yyyy", Locale.getDefault()) }
-
-    Card(
-        modifier  = Modifier.fillMaxWidth(),
-        shape     = RoundedCornerShape(14.dp),
-        elevation = CardDefaults.cardElevation(defaultElevation = 0.dp),
-        colors    = CardDefaults.cardColors(containerColor = Color.Transparent),
-        border    = BorderStroke(1.5.dp, Color(0xFF007A90).copy(alpha = 0.4f))
-    ) {
-        Row(modifier = Modifier.fillMaxWidth().padding(12.dp), verticalAlignment = Alignment.CenterVertically) {
-
-            Box(
-                modifier = Modifier
-                    .size(72.dp)
-                    .clip(RoundedCornerShape(10.dp))
-                    .background(Color(0xFFF5F5F5))
-                    .border(2.dp, Color(0xFF007A90).copy(alpha = 0.3f), RoundedCornerShape(10.dp)),
-                contentAlignment = Alignment.Center
-            ) {
-                val imageModel: Any? = item.imageUrl.takeIf { it.isNotBlank() } ?: item.image
-                val hasImage = imageModel != null && imageModel != Uri.EMPTY && imageModel.toString().isNotBlank()
-                if (hasImage) {
-                    AsyncImage(
-                        model = imageModel,
-                        contentDescription = item.title,
-                        modifier = Modifier.fillMaxSize().padding(6.dp),
-                        contentScale = ContentScale.Fit
-                    )
-                } else {
-                    Icon(Icons.Default.Checkroom, contentDescription = null, tint = Color.LightGray, modifier = Modifier.size(32.dp))
-                }
-            }
-
-            Spacer(Modifier.width(12.dp))
-
-
-            Column(modifier = Modifier.weight(1f)) {
-                Text(item.title.ifBlank { "Unnamed item" }, fontWeight = FontWeight.Bold, fontSize = 15.sp, maxLines = 1)
-                if (item.category.isNotBlank()) Text(item.category, fontSize = 12.sp, color = Color.Gray)
-                Spacer(Modifier.height(4.dp))
-                Surface(
-                    shape = RoundedCornerShape(8.dp),
-                    color = when {
-                        daysSince >= 365 -> Color(0xFFFFCDD2)
-                        daysSince >= 180 -> Color(0xFFFFE0B2)
-                        else             -> Color(0xFFFFF9C4)
-                    }
-                ) {
-                    Text(
-                        "Unworn ${daysSince}d · last ${sdf.format(item.lastWorn)}",
-                        modifier = Modifier.padding(horizontal = 8.dp, vertical = 3.dp),
-                        fontSize = 11.sp,
-                        color = Color(0xFF4E342E),
-                        fontWeight = FontWeight.Medium
-                    )
-                }
-            }
-
-
-            Column(
-                verticalArrangement = Arrangement.spacedBy(6.dp),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                modifier = Modifier.padding(start = 8.dp)
-            ) {
-                FilledTonalButton(
-                    onClick = onDonate,
-                    contentPadding = PaddingValues(horizontal = 10.dp, vertical = 4.dp),
-                    colors = ButtonDefaults.filledTonalButtonColors(
-                        containerColor = Color(0xFFE8F5E9),
-                        contentColor   = Color(0xFF2E7D32)
-                    )
-                ) {
-                    Icon(Icons.Default.CardGiftcard, contentDescription = null, modifier = Modifier.size(14.dp))
-                    Spacer(Modifier.width(4.dp))
-                    Text("Donate", fontSize = 12.sp)
-                }
-                OutlinedButton(
-                    onClick = onKeep,
-                    contentPadding = PaddingValues(horizontal = 10.dp, vertical = 4.dp)
-                ) {
-                    Icon(Icons.Default.Favorite, contentDescription = null, modifier = Modifier.size(14.dp))
-                    Spacer(Modifier.width(4.dp))
-                    Text("Keep", fontSize = 12.sp)
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun EmptyDonationState() {
-    Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-        Column(horizontalAlignment = Alignment.CenterHorizontally, verticalArrangement = Arrangement.spacedBy(12.dp)) {
-            Icon(Icons.Default.CheckCircle, contentDescription = null, tint = Color(0xFF2E7D32), modifier = Modifier.size(72.dp))
-            Text("Your wardrobe is clutter-free!", fontWeight = FontWeight.Bold, fontSize = 18.sp)
-            Text("No items inactive past your threshold.\nGreat job!", fontSize = 14.sp, color = Color.Gray, textAlign = TextAlign.Center)
         }
     }
 }
